requirements/base.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-environ==0.4.54
django-solo==2.2.0
requests==2.31.0
